third_party/blink/web_tests/transforms/transform-inherit-initial-unprefixed-expected.txt

Test that inherit and initial works on unprefixed transforms.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing inherit.
PASS computedStyle.transform is 'matrix(0.5, 0.866025, -0.866025, 0.5, 0, 0)'
PASS computedStyle.webkitTransform is 'matrix(0.5, 0.866025, -0.866025, 0.5, 0, 0)'
PASS computedStyle.transformOrigin is '100px 0px'
PASS computedStyle.webkitTransformOrigin is '100px 0px'
PASS computedStyle.transformStyle is 'preserve-3d'
PASS computedStyle.webkitTransformStyle is 'preserve-3d'
Testing initial.
PASS computedStyle.transform is 'none'
PASS computedStyle.webkitTransform is 'none'
PASS computedStyle.transformOrigin is '250px 0px'
PASS computedStyle.webkitTransformOrigin is '250px 0px'
PASS computedStyle.transformStyle is 'flat'
PASS computedStyle.webkitTransformStyle is 'flat'
PASS successfullyParsed is true

TEST COMPLETE
